requirements.txt
Django==5.2.3
djangorestframework==3.16.0
celery==5.5.3
redis==6.2.0
django-environ==0.11.2
psycopg2-binary==2.9.10
django-celery-beat==2.8.0
django-celery-results==2.5.1
django-cors-headers==4.7.0